Memory/RasterCanvas: make conversion consistent in ClipIncremental

+++ src/Screen/Memory/RasterCanvas.hpp
@@ -210,7 +210,7 @@ protected:
       if (code1 & CLIP_LEFT_EDGE) {
 	if ((y2 != y1) && (x1 != x2)) {
 	  const float m = float(y2 - y1) / float(x2 - x1);
-	  y1 -= x1 * m;
+	  y1 -= int(x1 * m);
 	  code1 = ClipEncodeY(y1);
 	} else {
 	  code1 &= ~CLIP_LEFT_EDGE;
@@ -237,7 +237,7 @@ protected:
       } else if (code1 & CLIP_TOP_EDGE) {
 	if ((y2 != y1) && (x1 != x2)) {
 	  const float m = float(x2 - x1) / float(y2 - y1);
-          x1 -= y1 * m;
+          x1 -= int(y1 * m);
 	  code1 = ClipEncodeX(x1);
 	} else {
 	  code1 &= ~CLIP_TOP_EDGE;
